chore: rename API as well (#12791)
