+++ docs/api/puppeteer.puppeteernode.md
@@ -63,7 +63,7 @@ Description
 </th></tr></thead>
 <tbody><tr><td>
 
-<span id="defaultproduct">defaultProduct</span>
+<span id="defaultbrowser">defaultBrowser</span>
 
 </td><td>
 
@@ -80,7 +80,7 @@ The name of the browser that will be launched by default. For `puppeteer`, this 
 </td></tr>
 <tr><td>
 
-<span id="lastlaunchedproduct">lastLaunchedProduct</span>
+<span id="lastlaunchedbrowser">lastLaunchedBrowser</span>
 
 </td><td>
 
@@ -111,7 +111,7 @@ string
 
 **Deprecated:**
 
-Do not use as this field as it does not take into account multiple browsers of different types. Use [defaultProduct](./puppeteer.puppeteernode.md#defaultproduct) or [lastLaunchedProduct](./puppeteer.puppeteernode.md#lastlaunchedproduct).
+Do not use as this field as it does not take into account multiple browsers of different types. Use [defaultBrowser](./puppeteer.puppeteernode.md#defaultbrowser) or [lastLaunchedBrowser](./puppeteer.puppeteernode.md#lastlaunchedbrowser).
 
 </td></tr>
 </tbody></table>
