+++ packages/puppeteer-core/src/node/FirefoxLauncher.ts
@@ -20,11 +20,11 @@ import {
 import {debugError} from '../common/util.js';
 import {assert} from '../util/assert.js';
 
+import {BrowserLauncher, type ResolvedLaunchArgs} from './BrowserLauncher.js';
 import type {
   BrowserLaunchArgumentOptions,
   PuppeteerNodeLaunchOptions,
 } from './LaunchOptions.js';
-import {BrowserLauncher, type ResolvedLaunchArgs} from './ProductLauncher.js';
 import type {PuppeteerNode} from './PuppeteerNode.js';
 import {rm} from './util/fs.js';
