+++ packages/puppeteer-core/src/node/PuppeteerNode.ts
@@ -22,6 +22,7 @@ import {type CommonPuppeteerSettings, Puppeteer} from '../common/Puppeteer.js';
 import type {SupportedBrowser} from '../common/SupportedBrowser.js';
 import {PUPPETEER_REVISIONS} from '../revisions.js';
 
+import type {BrowserLauncher} from './BrowserLauncher.js';
 import {ChromeLauncher} from './ChromeLauncher.js';
 import {FirefoxLauncher} from './FirefoxLauncher.js';
 import type {
@@ -29,7 +30,6 @@ import type {
   ChromeReleaseChannel,
   LaunchOptions,
 } from './LaunchOptions.js';
-import type {BrowserLauncher} from './ProductLauncher.js';
 
 /**
  * @public
@@ -79,7 +79,7 @@ export interface PuppeteerLaunchOptions
  */
 export class PuppeteerNode extends Puppeteer {
   #_launcher?: BrowserLauncher;
-  #lastLaunchedProduct?: SupportedBrowser;
+  #lastLaunchedBrowser?: SupportedBrowser;
 
   /**
    * @internal
@@ -168,8 +168,8 @@ export class PuppeteerNode extends Puppeteer {
    * @param options - Options to configure launching behavior.
    */
   launch(options: PuppeteerLaunchOptions = {}): Promise<Browser> {
-    const {browser: product = this.defaultProduct} = options;
-    this.#lastLaunchedProduct = product;
+    const {browser: product = this.defaultBrowser} = options;
+    this.#lastLaunchedBrowser = product;
     return this.#launcher.launch(options);
   }
 
@@ -179,11 +179,11 @@ export class PuppeteerNode extends Puppeteer {
   get #launcher(): BrowserLauncher {
     if (
       this.#_launcher &&
-      this.#_launcher.browser === this.lastLaunchedProduct
+      this.#_launcher.browser === this.lastLaunchedBrowser
     ) {
       return this.#_launcher;
     }
-    switch (this.lastLaunchedProduct) {
+    switch (this.lastLaunchedBrowser) {
       case 'chrome':
         this.defaultBrowserRevision = PUPPETEER_REVISIONS.chrome;
         this.#_launcher = new ChromeLauncher(this);
@@ -193,7 +193,7 @@ export class PuppeteerNode extends Puppeteer {
         this.#_launcher = new FirefoxLauncher(this);
         break;
       default:
-        throw new Error(`Unknown product: ${this.#lastLaunchedProduct}`);
+        throw new Error(`Unknown product: ${this.#lastLaunchedBrowser}`);
     }
     return this.#_launcher;
   }
@@ -229,8 +229,8 @@ export class PuppeteerNode extends Puppeteer {
   /**
    * The name of the browser that was last launched.
    */
-  get lastLaunchedProduct(): SupportedBrowser {
-    return this.#lastLaunchedProduct ?? this.defaultProduct;
+  get lastLaunchedBrowser(): SupportedBrowser {
+    return this.#lastLaunchedBrowser ?? this.defaultBrowser;
   }
 
   /**
@@ -238,15 +238,15 @@ export class PuppeteerNode extends Puppeteer {
    * `puppeteer`, this is influenced by your configuration. Otherwise, it's
    * `chrome`.
    */
-  get defaultProduct(): SupportedBrowser {
+  get defaultBrowser(): SupportedBrowser {
     return this.configuration.defaultBrowser ?? 'chrome';
   }
 
   /**
    * @deprecated Do not use as this field as it does not take into account
    * multiple browsers of different types. Use
-   * {@link PuppeteerNode.defaultProduct | defaultProduct} or
-   * {@link PuppeteerNode.lastLaunchedProduct | lastLaunchedProduct}.
+   * {@link PuppeteerNode.defaultBrowser | defaultBrowser} or
+   * {@link PuppeteerNode.lastLaunchedBrowser | lastLaunchedBrowser}.
    *
    * @returns The name of the browser that is under automation.
    */
